Rename files with kebab style
