+++ src/app/core/components/corecomponents.module.ts
@@ -11,9 +11,9 @@ import { CardComponent } from 'app/core/components/card/card.component';
 import { CopyButtonComponent } from 'app/core/components/copy-btn/copy-btn.component';
 import { Display } from 'app/core/components/display/display.component';
 import { PageComponent } from 'app/core/components/page/page.component';
+import { ViewChartAreaComponent } from 'app/core/components/view-chart-area/view-chart-area.component';
 import { ViewComponent } from 'app/core/components/view/view.component';
 import { ViewChartComponent } from 'app/core/components/viewchart/viewchart.component';
-import { ViewChartAreaComponent } from 'app/core/components/viewchartarea/viewchartarea.component';
 import { ViewChartDonutComponent } from 'app/core/components/viewchartdonut/viewchartdonut.component';
 import { ViewChartPieComponent } from 'app/core/components/viewchartpie/viewchartpie.component';
 import { ViewControlComponent } from 'app/core/components/viewcontrol/viewcontrol.component';
