+++ src/app/core/components/view-chart-area/view-chart-area.component.scss
